+++ src/test/scala/edu/ie3/simona/agent/em/EmAgentSpec.scala
@@ -67,7 +67,7 @@ class EmAgentSpec
 
   "A self-optimizing EM agent" should {
     "be initialized correctly and run through some activations" in {
-      val resultListener = TestProbe[ResultEvent]("ResultListener")
+      val resultProxy = TestProbe[ResultEvent]("ResultProxy")
       val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
       val emAgent = spawn(
@@ -78,7 +78,7 @@ class EmAgentSpec
           "PRIORITIZED",
           simulationStart,
           parent = Left(scheduler.ref),
-          listener = Iterable(resultListener.ref),
+          listener = resultProxy.ref,
         )
       )
 
@@ -116,7 +116,7 @@ class EmAgentSpec
       )
 
       // expect no results for init
-      resultListener.expectNoMessage()
+      resultProxy.expectNoMessage()
       // expect completion from EmAgent
       scheduler.expectMessage(
         Completion(emAgent, Some(0))
@@ -178,7 +178,7 @@ class EmAgentSpec
       )
 
       // expect correct results
-      resultListener.expectMessageType[FlexOptionsResultEvent] match {
+      resultProxy.expectMessageType[FlexOptionsResultEvent] match {
         case FlexOptionsResultEvent(flexResult) =>
           flexResult.getInputModel shouldBe emInput.getUuid
           flexResult.getTime shouldBe 0.toDateTime
@@ -187,7 +187,7 @@ class EmAgentSpec
           flexResult.getpMax() should equalWithTolerance(.006.asMegaWatt)
       }
 
-      resultListener.expectMessageType[ParticipantResultEvent] match {
+      resultProxy.expectMessageType[ParticipantResultEvent] match {
         case ParticipantResultEvent(emResult: EmResult) =>
           emResult.getInputModel shouldBe emInput.getUuid
           emResult.getTime shouldBe simulationStart
@@ -232,7 +232,7 @@ class EmAgentSpec
       emAgent ! FlexCompletion(modelUuid = evcsInput.getUuid)
 
       // expect correct results
-      resultListener.expectMessageType[FlexOptionsResultEvent] match {
+      resultProxy.expectMessageType[FlexOptionsResultEvent] match {
         case FlexOptionsResultEvent(flexResult) =>
           flexResult.getInputModel shouldBe emInput.getUuid
           flexResult.getTime shouldBe 300.toDateTime
@@ -241,7 +241,7 @@ class EmAgentSpec
           flexResult.getpMax() should equalWithTolerance(-.005.asMegaWatt)
       }
 
-      resultListener.expectMessageType[ParticipantResultEvent] match {
+      resultProxy.expectMessageType[ParticipantResultEvent] match {
         case ParticipantResultEvent(emResult: EmResult) =>
           emResult.getInputModel shouldBe emInput.getUuid
           emResult.getTime shouldBe 300.toDateTime
@@ -255,7 +255,7 @@ class EmAgentSpec
     }
 
     "revoke triggers correctly" in {
-      val resultListener = TestProbe[ResultEvent]("ResultListener")
+      val resultProxy = TestProbe[ResultEvent]("ResultProxy")
       val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
       val emAgent = spawn(
@@ -266,7 +266,7 @@ class EmAgentSpec
           "PRIORITIZED",
           simulationStart,
           parent = Left(scheduler.ref),
-          listener = Iterable(resultListener.ref),
+          listener = resultProxy.ref,
         )
       )
 
@@ -344,7 +344,7 @@ class EmAgentSpec
       )
 
       // expect correct results
-      resultListener.expectMessageType[FlexOptionsResultEvent] match {
+      resultProxy.expectMessageType[FlexOptionsResultEvent] match {
         case FlexOptionsResultEvent(flexResult) =>
           flexResult.getInputModel shouldBe emInput.getUuid
           flexResult.getTime shouldBe 0.toDateTime
@@ -353,7 +353,7 @@ class EmAgentSpec
           flexResult.getpMax() should equalWithTolerance(.006.asMegaWatt)
       }
 
-      resultListener.expectMessageType[ParticipantResultEvent] match {
+      resultProxy.expectMessageType[ParticipantResultEvent] match {
         case ParticipantResultEvent(emResult: EmResult) =>
           emResult.getInputModel shouldBe emInput.getUuid
           emResult.getTime shouldBe simulationStart
@@ -414,7 +414,7 @@ class EmAgentSpec
       )
 
       // expect correct results
-      resultListener.expectMessageType[FlexOptionsResultEvent] match {
+      resultProxy.expectMessageType[FlexOptionsResultEvent] match {
         case FlexOptionsResultEvent(flexResult) =>
           flexResult.getInputModel shouldBe emInput.getUuid
           flexResult.getTime shouldBe 300.toDateTime
@@ -423,7 +423,7 @@ class EmAgentSpec
           flexResult.getpMax() should equalWithTolerance(.008.asMegaWatt)
       }
 
-      resultListener.expectMessageType[ParticipantResultEvent] match {
+      resultProxy.expectMessageType[ParticipantResultEvent] match {
         case ParticipantResultEvent(emResult: EmResult) =>
           emResult.getInputModel shouldBe emInput.getUuid
           emResult.getTime shouldBe 300.toDateTime
@@ -438,7 +438,7 @@ class EmAgentSpec
     }
 
     "handle ChangingFlexOptions indicator correctly" in {
-      val resultListener = TestProbe[ResultEvent]("ResultListener")
+      val resultProxy = TestProbe[ResultEvent]("ResultProxy")
       val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
       val emAgent = spawn(
@@ -449,7 +449,7 @@ class EmAgentSpec
           "PRIORITIZED",
           simulationStart,
           parent = Left(scheduler.ref),
-          listener = Iterable(resultListener.ref),
+          listener = resultProxy.ref,
         )
       )
 
@@ -528,7 +528,7 @@ class EmAgentSpec
         requestAtTick = Some(600),
       )
 
-      resultListener.expectMessageType[FlexOptionsResultEvent] match {
+      resultProxy.expectMessageType[FlexOptionsResultEvent] match {
         case FlexOptionsResultEvent(flexResult) =>
           flexResult.getInputModel shouldBe emInput.getUuid
           flexResult.getTime shouldBe 0.toDateTime
@@ -537,7 +537,7 @@ class EmAgentSpec
           flexResult.getpMax() should equalWithTolerance(.006.asMegaWatt)
       }
 
-      resultListener.expectMessageType[ParticipantResultEvent] match {
+      resultProxy.expectMessageType[ParticipantResultEvent] match {
         case ParticipantResultEvent(emResult: EmResult) =>
           emResult.getInputModel shouldBe emInput.getUuid
           emResult.getTime shouldBe 0.toDateTime
@@ -607,7 +607,7 @@ class EmAgentSpec
       )
 
       // expect correct results
-      resultListener.expectMessageType[FlexOptionsResultEvent] match {
+      resultProxy.expectMessageType[FlexOptionsResultEvent] match {
         case FlexOptionsResultEvent(flexResult) =>
           flexResult.getInputModel shouldBe emInput.getUuid
           flexResult.getTime shouldBe 300.toDateTime
@@ -616,7 +616,7 @@ class EmAgentSpec
           flexResult.getpMax() should equalWithTolerance(.008.asMegaWatt)
       }
 
-      resultListener.expectMessageType[ParticipantResultEvent] match {
+      resultProxy.expectMessageType[ParticipantResultEvent] match {
         case ParticipantResultEvent(emResult: EmResult) =>
           emResult.getInputModel shouldBe emInput.getUuid
           emResult.getTime shouldBe 300.toDateTime
@@ -632,8 +632,7 @@ class EmAgentSpec
 
   "An EM-controlled EM agent" should {
     "be initialized correctly and run through some activations" in {
-      val resultListener = TestProbe[ResultEvent]("ResultListener")
-
+      val resultProxy = TestProbe[ResultEvent]("ResultProxy")
       val parentEmAgent = TestProbe[FlexResponse]("ParentEmAgent")
 
       val emAgent = spawn(
@@ -644,7 +643,7 @@ class EmAgentSpec
           "PRIORITIZED",
           simulationStart,
           parent = Right(parentEmAgent.ref),
-          listener = Iterable(resultListener.ref),
+          listener = resultProxy.ref,
         )
       )
 
@@ -685,7 +684,7 @@ class EmAgentSpec
       )
 
       // expect no results for init
-      resultListener.expectNoMessage()
+      resultProxy.expectNoMessage()
       // expect completion from EmAgent
       parentEmAgent.expectMessage(
         FlexCompletion(
@@ -723,7 +722,7 @@ class EmAgentSpec
         ),
       )
 
-      resultListener.expectMessageType[FlexOptionsResultEvent] match {
+      resultProxy.expectMessageType[FlexOptionsResultEvent] match {
         case FlexOptionsResultEvent(flexResult) =>
           flexResult.getInputModel shouldBe emInput.getUuid
           flexResult.getTime shouldBe 0.toDateTime
@@ -780,7 +779,7 @@ class EmAgentSpec
       )
 
       // expect correct results
-      resultListener.expectMessageType[ParticipantResultEvent] match {
+      resultProxy.expectMessageType[ParticipantResultEvent] match {
         case ParticipantResultEvent(emResult: EmResult) =>
           emResult.getInputModel shouldBe emInput.getUuid
           emResult.getTime shouldBe 0.toDateTime
@@ -830,7 +829,7 @@ class EmAgentSpec
       )
 
       // expect correct results
-      resultListener.expectMessageType[ParticipantResultEvent] match {
+      resultProxy.expectMessageType[ParticipantResultEvent] match {
         case ParticipantResultEvent(emResult: EmResult) =>
           emResult.getInputModel shouldBe emInput.getUuid
           emResult.getTime shouldBe 150.toDateTime